pyproject.toml
[tool.poetry]
name = "rctgcs"
version = "0.1.0"
description = ""
authors = ["Nathan Hui <nthui@eng.ucsd.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.9"
scipy = "^1.10.1"
utm = "^0.7.0"
appdirs = "^1.4.4"
xlwt = "^1.3.0"
ipython = "^8.11.0"
rctcomms = {git = "https://github.com/UCSD-E4E/radio_collar_tracker_comms.git"}
conda-lock = "^1.4.0"
charset-normalizer = "2.1.1"
colorama = "0.4.6"
idna = "3.4"
more-itertools = "9.1.0"
six = "1.16.0"
urllib3 = "1.26.14"
zipp = "3.15.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.1"
coverage = "^7.2.1"
pylint = "^2.16.2"
wheel = "^0.38.4"
PyAutoGUI = "^0.9.53"
pytest-qt = "^4.2.0"
pyqt5-stubs = "^5.15.6.0"

[tool.poetry.scripts]
RCTGcs = 'RctGcs.rctGCS:main'
droneSimulator = 'RctGcs.droneSimulator:main'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
